Fixed the Simulate function / prob of successes
Changed probability of successes from 0.5 which was giving nearly the same values for every signal detection to a formula giving hit rate and false alarm rate using criterion and dprime values. The hits used hit rate, misses 1 - hit rate, and FA/CR the same but with FA rate. Now when using plot roc on the simulated data the curve looks much more typical and changes based on the dprime value inputted. Only issue now is the code smells of having the ROC Loss function rewritten in fit_roc but I cannot at the moment find a solution to fixing that without it saying it needs us to define a.